backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.3.2
python-dateutil>=2.8.2
scikit-learn==1.6.1
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
APScheduler>=3.10.0
lightgbm>=4.0.0

